+++ services/google-data-api/src/utils/build-youtube-url.js
@@ -2,4 +2,4 @@ const { join } = require('path');
 
 const { BASE_URL, YOUTUBE_ENDPOINT } = require('../constants');
 
-module.exports = endpoint => `${BASE_URL}${join(YOUTUBE_ENDPOINT, endpoint)}`;
+module.exports = endpoint => `${BASE_URL}${join('/', YOUTUBE_ENDPOINT, endpoint)}`;
